'Working' Implementation of Dates on server/javscript side
This needs to be changed to remove any chance that the update will
change the created or last_updated values in the db. It should
only change the due_date if there is a value. In general, the
update should only change things if they are sent in.

There may also be some error with waiting for disc read/write, but
that may also be my machine.